Create hackerDictionary.java
This problem will take inputs as names and phone numbers. This will be stored in a dictionary. Then, user can search for a phone number by the name which is key in our case. If the name is found in dictionary, it prints our the name and phone numbers, if not, then it print "Not Found".
Sample Input:
3
sam
99912222
tom
11122222
harry
12299933
sam
edward
harry

Sample Output:
sam=99912222
Not found
harry=12299933